requirement.txt
google-generativeai
feedparser